refactored route logic to a helper file to keep server.js clean/legible
